Fixed project structure
