+++ src/pages/Main/MainScreen.js
@@ -1,0 +1,17 @@
+import React, { useState } from 'react';
+import Energys from './components/Energys';
+import Rounds from './components/Rounds';
+
+export function MainScreen() {
+  const [energy, setEnergy] = useState(3)
+  const [round, setRound] = useState(1)
+
+  return (
+    <div className="w-full h-full bg-gray-900 flex justify-center items-center">
+      <div className="w-80 h-80 bg-white rounded-2xl flex flex-col items-center">
+        <Rounds setEnergy={setEnergy} round={round} setRound={setRound} />
+        <Energys energy={energy} setEnergy={setEnergy} />
+      </div>
+    </div>
+  );
+}
